Add initial CoDiOS emulator support, and build improvements
I removed .cargo/config, as it is no longer necessary with the build
improvements.

We now have two features: `emulator`, and `firmware`. These are pretty
much self-explantory, but essentially, when you build targeting
`aarch64` OR `x86_64`, and specify the `emulator` feature, this will
build the CoDiOS emulator.

You should not build the firmware using the `aarch64` OR `x86_64`
targets, as it's specifically designed for the CoDi STM32 chip.

When building using the `thumbv7em-none-eabihf` target, and the
`firmware` feature, Cargo will build the firmware for the STM32 chip,
and you can then flash it using the new CoDi bootloader.

The new emulator is available, and currently very barebones.

I have also updated Cargo.lock, and moved the `bin` folder to
`src/bin`, in line with Rust guidelines.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>